Fix invalid handling of llvm MemoryBuffer in DiagnosticMain.
When 'Buf' is passed to clang in addRemappedFile then clang will destroy it when the compiler instance is destroyed.
So release ownership of the pointer immediately.
Otherwise if the call to Parse fails then we destroy both the compiler instance (indirectly destroying the MemoryBuffer) and destroy the memory buffer again via the reference to it from 'Buf' which isn't cleared yet.
This can cause a segmentation fault.